app and public now submit combined output to coveralls.io

+++ coverage.sh
@@ -4,14 +4,28 @@
 # script based on linked document:
 # https://www.npmjs.com/package/mocha-lcov-reporter
 
-rm -rf coverage
-rm -rf app-cov
- 
-node_modules/.bin/jscover --exclude=tests app app-cov # exclude tests from instrumentation ...
-cp -r app/tests app-cov # ... replace tests so mocha can work
+[ -d coverage ] && rm -rf coverage
+mkdir coverage
 
+# NOTE getting coverage for karma requires less hoop jumping
+node_modules/.bin/karma start karma.coveralls.conf.js
+
+# NOTE getting coverage for mocha is not so simple
+# build instrumented code and backup original before overwriting
+[ -d app-orig ] && rm -rf app-orig
+node_modules/.bin/jscover app coverage/app
 mv app app-orig
-mv app-cov app
-node_modules/.bin/mocha app/tests -R mocha-lcov-reporter | node_modules/coveralls/bin/coveralls.js app
-rm -rf app
+mv coverage/app app
+
+mkdir coverage/app
+#node_modules/.bin/mocha -R mocha-lcov-reporter app/tests > coverage/app/lcov.info
+JS_COV=1 node_modules/.bin/mocha -R mocha-lcov-reporter app/tests > coverage/app/lcov.info.temp
+# fix the paths (https://www.npmjs.com/package/mocha-lcov-reporter)
+sed 's,SF:,SF:app/,' coverage/app/lcov.info.temp > coverage/app/lcov.info
+
+# now both client and server are complete, send reports to coveralls together
+cat coverage/**/lcov.info | node_modules/coveralls/bin/coveralls.js
+
+mv app coverage/app
 mv app-orig app
+
